Refactors simultaneous downloads / dark mode to directly reference VB Settings.
It was looking like the previous approach would duplicate a lot of the existing Settings framework with the downside of not detecting changes if anything modified Settings directly instead of through ProgramSettings.
Using Settings as the source of truth is better, even if this does make it little more than a wrapper over Settings.