Adds ability for DateRange tag to get every other attribute that didn't already match.
This is because it can contain custom client attributes or SCTE-35 data.
Ideally this would be preserved if the playlist is written as a file again.

There may be additional tags (starting with a "#") present in an HLS file but these can be vendor-specific and look out of scope for this project.