Creates basic structure of playlist processing / output targets.
Parsing a playlist first goes through an IPlaylistProcessor that parses the playlist into an object and chooses the correct items to download.

Control is handed to the ITarget that can retrieve whatever it wants from the processor. This can either be a list of segments or the literal playlist to let FFmpeg handle it.