Big rewrite of HLS playlist parsing.
All tag objects have been renamed to a TagParser with a common superclass. Each parser is an object that takes a stream and can add the appropriate properties to a playlist object. Parsing a playlist now gets the tag, retrieves the appropriate parser, and has that parser append to the playlist.

This replaces the duplicated tag names, long select statement, and inconsistent handling of tags that define a boolean value, an entire object, or define a segment (when combined with other tags).