Add a framework for a slightly different downloading algorithm.
It makes the download portion more independent of the playback selection. The subtitles and video tracks can be defined anywhere (even though HLS can include subtitles, there is no guarantee the subtitles aren't in a separate file). The responsibility of the client is to produce a list of playbacks, and the downloader is responsible for getting all media in the playback.

- A playback is the set of all media that can definitely be played together, but it doesn't have to if all the tracks in it weren't requested to be downloaded.
- An individual episode may have different playbacks if the requested tracks don't all come from the same playlist.
- A media object contains the media in a usable state. For example, it could be a URL to a subtitle file or a parsed HLS media playlist.
- Media is tagged with the kind of data it contains (subtitle, video, audio). This is for situations where a video stream contains embedded audio.