Adds download progress events.
Events are needed in the downloader and the ffmepg adapter.

This will allow the downloaders to report how much of the subtitle or audio has completed.
Also changes the DownloadPlaybacks method signature to accept a single selection. A selection already contains everything that is supposed to be downloaded, so the list doesn't make sense here.