Bug fixes around ffmpeg arguments in mp4postprocessor.
Only add a file to the ffmpeg input files list if it is being selected.
Use correct input file number even when the file is skipped (such as not muxing subtitles).
Have Mp4Postprocessor#ProcessFile return whether the file was used so that the outer loop can maintain the correct input file index.
Return the files produced so that the next stage knows what files to work with.
Convert ProcessInputs to an async method so ffmpeg can be awaited.
Remove episode parameter from ProcessInputs as it is no longer needed.