Refactors useragent and cookies into FfmpegArguments.
They should have been here in the first place, but they were placed in FfmpegAdapter thinking they couldn't cause any problems.
In reality, it can cause a remux of a local file to fail because user_agent isn't allowed in that context.
And the nonzero error code was being swallowed so it wasn't obvious what was happening. Therefore, also changed the taskCompletionSource to set an exception status if the exit code is nonzero.