Uses program settings to build correct output preferences.
Changes subtitle language naming to allow skipping only when a single language is being downloaded. The reason is that subtitle order is not deterministic, so if downloading multiple episodes, it may name files inconsistently. The spirit of this seems to be allowing you to have a shorter filename when only downloading one language.