Adds download time estimation and properly reports progress for the entire download.
Estimation is a simple wrapper around a Stopwatch object, and it tracks the time for the current stage. This keeps timing accurate when switching between two operations (from download to re-encoding, for example).

Downloading has been converted into a standard set of objects that take an input and produce an output. It is called a pipeline, but it isn't built up-front and executed at the end like a true pipeline. This is just to reduce the amount of boilerplate and increase readability when repeatedly reporting progress.