Remove console window from Main.
The TextBoxTraceListener caused a crash, with the stack trace mentioning an inability to create a window handle. I think creating a new Timer each time the program wrote to the console exhausted the number of window handles.

The console is only useful to developers, who have the console in the IDE so this isn't needed.