Delete legacy FunimationDownloader.
This was originally broken out of Main.vb to make it more manageable.
It has since been replaced by the client code now in the SiteApi project. As all of the site-specific logic has been implemented, it is no longer needed. Any TODOs addressing general program behavior have been removed as obsolete or moved to an appropriate location for future work.