Allow crunchyroll client to send authenticated requests.
There is still some work here. The code is pretty messy and hard to follow. Crunchyroll, unlike funimation, requires cookies to send any requests and needs an initialization step. To make this a little cleaner, there is a new interactive cookie interface that allows the authenticator to request cookies for a site. The browser implements this by navigating to that URI, waiting for the page to load completely, and gettting cookies as normal. This is a little better than forcibly opening the browser when adding an episode, as was done before.